Atsumaru (EN): fix search 404, implement TriState filters and metadata
- Migrate search and filter to new Typesense API endpoint.
- Update MangaDto to handle polymorphic author and tag fields.
- Correctly build filter_by query parameters, ensuring mandatory filters to avoid 404.
- Implement TriState genre filters (Include/Exclude) matching website behavior.
- Update sorting options to include Views and use field:order format.
- Enhance metadata with Rating, Synopsis, and Alternative Names in the requested format.
- Consolidate filter code into Filters.kt and cleanup unused files.
- Bump version to 16.
